removes piston, installs minifb, rendering now is much faster, implements idiomatic rust in few places

+++ src/fluid_simulation/collision_manager.rs
@@ -7,7 +7,7 @@ pub struct CollisionManager {
 }
 
 impl CollisionManager {
-  pub fn new(box_dimensions: [i32; 2]) -> Self {
+  pub fn new(box_dimensions: [usize; 2]) -> Self {
     CollisionManager {
       box_width: box_dimensions[0] as f32,
       box_height: box_dimensions[1] as f32,
